adafruit_bme680.py: Fix compensation of pressure values
Compared the ot reference code, one expression in the lengthy calculation was split in two lines. 
Since a previous interim value was used in the expression twice and replaced by the first 
part of the split expression, the overall result was different to the reference code. The resulting
pressure was not that different, 10 hPa at 1000hPa absolute value, but clearly outside of the
spec with 1hPa absolute pressure error.